# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API da loja de chapéus Hat Store.",
        "title": "Hat Store API",
        "contact": {},
        "version": "1.0"
    },
    "host": "hatstore-prd.fly.dev",
    "basePath": "/api/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Loga um usuário com email e senha e retorna um token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Autentica um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de desenvolvimento",
                        "name": "X-Dev-UUID",
                        "in": "header"
                    },
                    {
                        "description": "Credenciais de Login",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Email ou senha inválidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Muitas tentativas de login",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Cria uma nova conta de usuário com email e senha",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Registra um novo usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de desenvolvimento",
                        "name": "X-Dev-UUID",
                        "in": "header"
                    },
                    {
                        "description": "Credenciais de Registro",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso!",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Usuário já existe",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "rememberMe": {
                    "type": "boolean"
                }
            }
        },
...
        "models.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}